Movements fixed
Defined the methods merge_and replace that takes a row or column and the direction to iterate the array

This method is called for every row or column of the board.

For example, for the moveUp method, we have to check every column from top to down to check every possible merge and place every value to the highest position as possible.