boardState and random_generator changes
board State now has a father attribute for the algorithm expansion and the functions g(), h(), f() and getSuccessors() have been added.

The random_generator has been changed to the numpy.random.default_rng(seed) as it allows for instantiation of multiple generators that share the sate of other generator. This way the random_generator of the successors will all produce the same result.

*Bug*: When using the fuction insert_random_number() sometimes the numpy.isin() function throws the exception: "RecursionError: maximum recursion depth exceeded"