Show A* results  on screen
Created a new view class, InfoView that receives the number of expanded nodes by A* and the algorithm execution time.

It draws that info on upper left corner of the screen, only on the auto mode.

Had to change the output of the AStar class algorithm method to return those variables.

Also changed boardView to receive the screen as a class variable, so that we don't need to pass it on the paint method every time