tiles_merging fix
Fixed a problem with the logic of merging tiles. Previously once the fuction found one merge in a row/colum it stopped cheking.

The iteration process is now diferent. Previously it deleted elements from the array that was being iterated over, which lead to out of bounds exceptions if you wanted to keep searching for multiple mergings