Node class and abstract cost class
In order to keep the game logic and the a* search independent, I created defined a Node class that contains all the expected behavior.

Also, the different g and h methods where confusing and misleading so I decided to apply the Strategy design pattern to keep it clean and simple. All implementations of g and h must inherit an abstract class.
All nodes share an instance of an AbstractCost superclass.

I moved the Rukai's costs to a superclass called rukai_costs.py.

All of that means that I removed all those things from the BoardState class. I kept all the different heuristics commented.

Made the random number generator be an static variable, so that there is only one. I don't know if it 's because of this change, but now the results are reproducible.
Also moved the SEED to the config file.