Fixes and refactor
Fixes: fixed previous rng refactor that made all succesors use the same rng so all of them would generate different numbers. Now the state of the rng is the same for all succesors of a node again.

Refactor: Divided cost and heuristic strategies in diferent files for organization and cleanness. Now each strategy has a single responsibility, following the Single Responsability Principle (before it had to calculate cost and heuristics with the same strategy). This allows to have diferents costs/heuristics while maintaining the heuristic/cost more easily.

Also refactored the node class. Now is an abstract class representing a node of an algorithm, and has abstract methods for g(), h(), and get_successors(). f() is not abstrac as it is just g()+h(). The board_state extends this class and implements said methods using the cost and heuristic strategies or defining how succesors are made.